Add support for arguments to the `packer_fmt` hook
This just pulls Packer arguments out to separate them from the files
that are passed in by pre-commit. Note that this is a minimal effort
implementation so arguments with values should use the `-arg=value`
form.

This also breaks the `-check` argument out into the overwritable `args`
value for the hook.