Support file paths with a leading `-`
This covers the more uncommon case where a file path provided to the
hooks has a leading `-`. This will prevent such a file path from being
treated as an argument, but it does mean that if a file with the same
name as an argument exists it will not be used as such.