Use the default behavior of `packer fmt`
Instead of providing default arguments to the `packer_fmt` hook we
instead use the default behavior of the `packer fmt` command. This
will result in files being written to if formatting is needed instead
of just warning that a file is improperly formatted. An important note
is that `packer fmt` returns 0 even if changing a file with the default
arguments. The existing exit code check is left in place in case the
arguments for the hook are changed.